Correct minor typos and restructure tests
