+++ src/test/java/seedu/address/logic/commands/customer/EditCustomerDescriptorTest.java
@@ -13,7 +13,7 @@ import static seedu.address.logic.commands.CommandTestUtil.VALID_TAG_HUSBAND;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.logic.commands.customer.EditCustomerCommand.EditPersonDescriptor;
-import seedu.address.testutil.EditPersonDescriptorBuilder;
+import seedu.address.testutil.customer.EditPersonDescriptorBuilder;
 
 public class EditCustomerDescriptorTest {
