+++ src/test/java/seedu/address/model/customer/NameContainsKeywordsPredicateTest.java
@@ -9,7 +9,7 @@ import java.util.List;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.testutil.PersonBuilder;
+import seedu.address.testutil.customer.PersonBuilder;
 
 public class NameContainsKeywordsPredicateTest {
