+++ src/test/java/seedu/address/storage/JsonSerializableAddressBookTest.java
@@ -11,7 +11,7 @@ import org.junit.jupiter.api.Test;
 import seedu.address.commons.exceptions.IllegalValueException;
 import seedu.address.commons.util.JsonUtil;
 import seedu.address.model.InventorySystem;
-import seedu.address.testutil.TypicalPersons;
+import seedu.address.testutil.customer.TypicalPersons;
 
 public class JsonSerializableAddressBookTest {
