+++ src/test/java/seedu/address/testutil/customer/EditPersonDescriptorBuilder.java
@@ -1,4 +1,4 @@
-package seedu.address.testutil;
+package seedu.address.testutil.customer;
 
 import java.util.Set;
 import java.util.stream.Collectors;
